Changed msb/lsb constructor
Changed Java compatible constructor to a freestanding function. This is in relation to #29 and the problem that defining constructors for our class is something that hinders us in general from being compatible with different kinds of native implementations. The goal is that we can define the `Uuid` class as a `typealias` in Java but we cannot, for instance, require a `Uuid(data: ByteArray)` constructor since it is already defined and private. There might be the possibility to overwrite this somehow with a freestanding function of the same name but I believe that it is simpler to just go for freestanding functions for everything and stick with the proper naming for them. Should make for a cleaner API overall.

I also improved the CHANGELOG a little further and please note that this change is not breaking because we never released the `Uuid(msb: Long, lsb: Long)` constructor.